1214: Increase AM maxHttpHeaderSize to 16k
AM currently does not config the Tomcat maxHttpHeaderSize, the default of 8k is used.

Issues have been observed with OAuth2.0 /authorize calls passing headers larger than 8k.

Applying the maxHttpHeaderSize configuration to the Tomcat server.xml. Using sed to append the configuration to the end of the Connector elements in the configuration file.

Note: FIDC sets maxHttpHeaderSize as 16k, this change bring CDK deployments in line with this.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1214